+++ api/src/main/java/io/strimzi/api/kafka/model/common/authentication/KafkaClientAuthentication.java
@@ -27,6 +27,7 @@ import java.util.Map;
     @JsonSubTypes.Type(name = KafkaClientAuthenticationScramSha512.TYPE_SCRAM_SHA_512, value = KafkaClientAuthenticationScramSha512.class),
     @JsonSubTypes.Type(name = KafkaClientAuthenticationPlain.TYPE_PLAIN, value = KafkaClientAuthenticationPlain.class),
     @JsonSubTypes.Type(name = KafkaClientAuthenticationOAuth.TYPE_OAUTH, value = KafkaClientAuthenticationOAuth.class),
+    @JsonSubTypes.Type(name = KafkaClientAuthenticationCustom.TYPE_CUSTOM, value = KafkaClientAuthenticationCustom.class),
 })
 @JsonInclude(JsonInclude.Include.NON_NULL)
 @EqualsAndHashCode
@@ -34,13 +35,13 @@ import java.util.Map;
 public abstract class KafkaClientAuthentication implements UnknownPropertyPreserving {
     private Map<String, Object> additionalProperties;
 
-    @Description("Authentication type. " +
-            "Currently the supported types are `tls`, `scram-sha-256`, `scram-sha-512`, `plain`, and 'oauth'. " +
-            "`scram-sha-256` and `scram-sha-512` types use SASL SCRAM-SHA-256 and SASL SCRAM-SHA-512 Authentication, respectively. " +
-            "`plain` type uses SASL PLAIN Authentication. " +
-            "`oauth` type uses SASL OAUTHBEARER Authentication. " +
-            "The `tls` type uses TLS Client Authentication. " +
-            "The `tls` type is supported only over TLS connections.")
+    @Description("Specifies the authentication type. " +
+            "Supported types are `tls`, `scram-sha-256`, `scram-sha-512`, `plain`, 'oauth', and `custom`. " +
+            "`tls` uses TLS client authentication and is supported only over TLS connections. " +
+            "`scram-sha-256` and `scram-sha-512` use SASL SCRAM-SHA-256 and SASL SCRAM-SHA-512 authentication, respectively. " +
+            "`plain` uses SASL PLAIN authentication. " +
+            "`oauth` uses SASL OAUTHBEARER authentication. " +
+            "`custom` allows you to configure a custom authentication mechanism.")
     public abstract String getType();
 
     @Override